refactor(): restructure unit tests
